requirements.txt
streamlit==1.50.0
tensorflow==2.20.0
joblib==1.5.2
numpy==2.2.6
Pillow==11.3.0
scikit-learn==1.6.1
matplotlib==3.10.7
seaborn==0.13.2
